requirements.txt
streamlit
python-dotenv
PyMuPDF
langchain
google-generativeai
langchain-google-genai